requirements.txt
customtkinter
Pillow
packaging
yt-dlp
